Corrected type casting in time/parser.hpp
Fixed time quantities being cast as int instead of long int

closes #173

+++ include/boost/astronomy/time/parser.hpp
@@ -29,9 +29,9 @@ public:
       dh = d;
 
       time = dh;
-      hours = (int)time;
+      hours = (long int)time;
       minutesRemainder = (time - hours) * 60;
-      minutes = (int)minutesRemainder;
+      minutes = (long int)minutesRemainder;
       secondsRemainder = (minutesRemainder - minutes) * 60;
       seconds = secondsRemainder;
     }
